doc: fix minor typographical error in DEP0079 text
Change `backwards compatibility` to `backward compatibility`.

(It's confusing because "backwards compatible" is acceptable because
"backwards" can be used as an adverb like that. However, as an
adjective, as in "backward compatibility", only "backward" will do.
Easiest solution: Always use "backward" because it is OK in both cases.
This is all compounded by the US vs UK English thing. US English tends
to favor "backward", and we standardize on US English, so that's another
point in favor of "backward" over "backwards" in this context.)

PR-URL: https://github.com/nodejs/node/pull/20519
Reviewed-By: Vse Mozhet Byt <vsemozhetbyt@gmail.com>
Reviewed-By: Refael Ackermann <refack@gmail.com>
Reviewed-By: Luigi Pinca <luigipinca@gmail.com>
Reviewed-By: Ruben Bridgewater <ruben@bridgewater.de>
Reviewed-By: James M Snell <jasnell@gmail.com>
Reviewed-By: Trivikram Kamat <trivikr.dev@gmail.com>